Add `add_Let` function.
It takes an expression, binds it to a fresh variable and evaluates the
body, providing it with the name (Id.t) that was just generated.
An example how it works is given by the k-normalization of Syntax.Add.
The idea behind this function is to use it for Syntax.App however I am
failing at making this work.