Improve `add_let`
If the expression is alraedy a variable, no new variable is introduced.
This change is necessary for the closure-conversion because otherwise
every function call will be need to be a closure.
Previously
`print_int 4`
was transformed into
```
let t1 = print_int in
let t2 = 4 in
(t1 t4)
```
With this change:
```
let t2 = 4 in
print_int t2