Frontend: Fix naked floats *again*
The translation from Closure.Float to ASML requires the introduction of
let-bindings which was previously not possible since Asml.t and Asml.exp
are different types. That is `closure_to_exp` could not return let
statements.
We moved the conversion of floats from `closure_to_exp` to `closure_to_t`
and adapted the nested let reduction code for Knorm.Lets to Asml.Let.